+++ Src/MoneyFox.Business/StatisticDataProvider/CategorySummaryDataProvider.cs
@@ -33,8 +33,8 @@ namespace MoneyFox.Business.StatisticDataProvider
                         .GetList()
                         .Where(x => (x.Date.Date >= startDate.Date) && (x.Date.Date <= endDate.Date))
                         .Where(x => x.CategoryId == category.Id)
-                        .Where(x => x.Type != (int) PaymentType.Transfer)
-                        .Sum(x => x.Type == (int) PaymentType.Expense
+                        .Where(x => x.Type != PaymentType.Transfer)
+                        .Sum(x => x.Type == PaymentType.Expense
                             ? -x.Amount
                             : x.Amount)
                 });
